+++ internal/driver/registry_factory.go
@@ -82,6 +82,7 @@ func NewDefaultRegistry(ctx context.Context, flags *pflag.FlagSet, withoutNetwor
 		ctxer:                     options.Contextualizer(),
 		defaultUnaryInterceptors:  options.GRPCUnaryInterceptors(),
 		defaultStreamInterceptors: options.GRPCStreamInterceptors(),
+		defaultGRPCServerOptions:  options.GRPCServerOptions(),
 		defaultHttpMiddlewares:    options.HTTPMiddlewares(),
 		extraMigrations:           options.ExtraMigrations(),
 		defaultMigrationOptions:   options.MigrationOptions(),
@@ -94,6 +95,12 @@ func NewDefaultRegistry(ctx context.Context, flags *pflag.FlagSet, withoutNetwor
 	}
 	if err := init(ctx); err != nil {
 		return nil, errors.Wrap(err, "unable to initialize service registry")
+	}
+
+	if inspect := options.Inspect(); inspect != nil {
+		if err := inspect(r.Persister().Connection(ctx)); err != nil {
+			return nil, errors.Wrap(err, "inspect")
+		}
 	}
 
 	return r, nil